Draft use JSON keymap and existing functionality with custom keycodes
