+++ keyboards/sofle/keymaps/rgb_default/rules.mk
@@ -5,3 +5,4 @@ RGBLIGHT_ENABLE = no
 ENCODER_ENABLE = yes
 LTO_ENABLE = yes
 OLED_ENABLE = no
+SRC += base_keymap.c
