+++ keyboards/sofle/sofle.h
@@ -1,0 +1,11 @@
+
+enum custom_keycodes {
+    KC_QWERTY = SAFE_RANGE,
+    KC_QWERTY_MAC,
+    KC_COLEMAK,
+	  KC_COLEMAKDH,
+    KC_LOWER,
+    KC_RAISE,
+    KC_ADJUST,
+    KC_D_MUTE
+};
